+++ Microsoft.Azure.Cosmos/src/Json/JsonTextParser.cs
@@ -6,6 +6,7 @@ namespace Microsoft.Azure.Cosmos.Json
     using System;
     using System.Buffers.Text;
     using System.Text;
+    using Microsoft.Azure.Cosmos.Query.Core;
 
     /// <summary>
     /// Common utility class for JsonTextReader and JsonTextNavigator.
@@ -99,7 +100,7 @@ namespace Microsoft.Azure.Cosmos.Json
                 throw new ArgumentOutOfRangeException($"Tried to read {value} as an {typeof(long).FullName}");
             }
 
-            return value;
+            return (long)value;
         }
 
         public static uint GetUInt32Value(ReadOnlySpan<byte> intToken)
